+++ modules/utils/utils_windows.go
@@ -29,8 +29,10 @@ func ExecShell(ctx context.Context, command string) (string, error)  {
     }()
     select {
         case <- ctx.Done():
-            exec.Command("taskkill", "/F", "/T", "/PID", strconv.Itoa(cmd.Process.Pid)).Run()
-            cmd.Process.Kill()
+            if cmd.Process.Pid > 0 {
+                exec.Command("taskkill", "/F", "/T", "/PID", strconv.Itoa(cmd.Process.Pid)).Run()
+                cmd.Process.Kill()
+            }
             return "", errors.New("timeout killed")
         case result := <- resultChan:
             return ConvertEncoding(result.output), result.err
